Crash fix. Prevent current line out-of-bound crash (Lines[n].Current > High(Lines)
The previous repeat loop was incrementing the line counter if the audio position is after the last beat. Out-of-bound crash occurred due to the game due to accessing the Lines array with a bigger index than than the High value (Length).

The while checks only the previous last item (High-1; Length-2) as it checks if the current beat is after the end of that line. In that case, it increases the index to High (Length-1) which will properly cover all lines in that array.