VERSION file stating the development version for the codebase
This version will define what kind of version the code belongs to - when dumped. This version is for development not for release. This version string is compatible to the semantic versioning 2.0. The appendix ("+") can be seen as meta data.

This specific appendix ("+") is used to differentiate the working files from the release (and its source code). Once a build will be released, the VERSION file will be updated to the precedence of the release name, and committed. Right after that point, it will be changed to that new VERSION string with the appendix. This will result that there's only 1 specific code dump with the matching version string of the release (its precedence). This commit is generally tagged as release, it can have other release related files committed.

If a release does have extra tags, such as alpha, beta and all that, it can be included in the release version file (check semver.org). An additional plain text name should be entered into a new line.

That said, if a new release is called "2.4.5-beta", this will be the committed content of the VERSION file, the master branch will get a new VERSION commit with the string "2.4.5+". This will state an ongoing development from the version 2.4.5. It doesn't matter if it's the same as a possible alpha release of that version. The code changes went in to the main branch (master). Branching the tagged release, that branch should update its version file matching the same string (including the pre-release identifier). This should be used if fixes to a specific release are created.

If an internal/public dev pre-release is published, it should have that VERSION file string from that codebase including in the name, additionally it can have a readable name right after "+". Other meta data should be added according the recommendation on semver.org. All these builds (but not the release) should have a readable metadata suffix in order to differentiate test builds build from each other (preferably a date is understandable across all user bases).

All in all, this should help providing pre-releases, test builds, fork builds and not mix it with the release cycle, and its versioning.