Hack fix. Multiple screen resolution wasn't stored correctly
By having a setup with multiple screens (Screens=2), the resolution wasn't properly retrieved by the getter Ini.GetResolution, and stored with Ini.SetResolution. This resulted into a bad indexed resolution when quitting the after having changed the resolution in a splitscreen and starting the program again.

This is a hackfix as the splitscreen is calculated in the graphics unit/code without properly state in which mode it is running. This should be centralized in graphics/ini and accessible.

In response to ...

"fixed index out of bounds exception for IResolution"
a6aae9631b635607afa95cd06d46dbf9348537ee

Revert "fixed index out of bounds exception for IResolution"
cb1056b321b66befdc2934cacdb5d8d6286a5fd3