Prevent bad config values incorrectly pointing into array
In most/all cases, the config values are read by GetArrayIndex with the DefaultArray passed. Sometimes the array is directly accessed by passing the default value. This is likely not the problem as this would be noticed once you start the game (and the ini is read). However, with a bad configuration, the default value (and thus the safe one) is not used at all.

Take a look at "Screens=3". The config value is stored and read like:
Screens := GetArrayIndex(IScreens, IniFile.ReadString('Graphics', 'Screens', IScreens[0]));

With the IniFile.ReadString(..) expression, the return value is not IScreens[0] (which is '1') but the properly read config value from the file ('3'). It will return -1, as the value is not found.

This value is not part of the possible values of that array.
IScreens:          array[0..1] of UTF8String  = ('1', '2');

Strangely, accessing static array with an index larger or lower their bounding will result into some unwanted value (from the memory). This is due to RangeChecks being set off (-$R-).

IScreens[Screens]

GetArrayIndex returns -1, and thus the Screens value is somehow not the desired but also not the default value.

The config value for Screens will be written _further down the line_:
IniFile.WriteString('Graphics', 'Screens', IScreens[Screens]);

Checking the config, it will be:
Screens=200

The problematic case is something like. It results into a direct crash.
MaxFramerate=61