Party mode selection screen
Instead of having a party menu with a slider to select the party mode and change specific sliders accordingly (and eventually switch to another screen) this new screen route the user to a specific mode's options.

This is a WIP screen and will point to the specific screens instead of ScreenPartyOptions.