prefer PTS over DTS and use simpler reordering code
If FFmpeg is able to tell us the PTS of a frame, use that value.
It also knows better than us which packet caused the current frame
to be returned and what its DTS was.

The old method of passing the PTS through the decoder relied on
the order of the frame allocation. With more recent (well, 8 years
old) FFmpeg versions, there is a reordered_opaque field in the
codec context that was designed to transport the PTS through
frame reordering. Even more recent (6 years) FFmpeg versions already
do the PTS reordering so that we just have to read the value from
the frame.