Allow additional linker flags
Introduce new variable PLDFLAGS in Makefile.in which can be used to
append additional flags to the linker. This variable will be used
later by Debhelper when building Debian packages.

The PLINKFLAGS cannot be used for this task because it is defined inside
Makefile.in and thus nothing can be appended to it. When appending
anything to PLDFLAGS, you have to use the Pascal Compiler syntax.