add workaround for FFmpeg bug #6560
Our av_seek_frame provokes a bug in the MOV/MP4 demuxer that causes it to
use wrong PTS-DTS deltas for the frames in the second and subsequent
fragments.

Prefer the DTS for files using this demuxer as it was done for all file
types before ab5efd2ddbf6fc7dc32b34215a65d29225208ad6.