UVideo: make RGB frame stride divisible by 32
Libswscale has some undocumented requirements regarding the stride of
the output buffer. F.ex. on x86 it might drop pixels on the right if
the stride is smaller than the width rounded up to a multiple of 8.

32 is the alignment used by libavfilter's ff_default_get_video_buffer
and the muxing.c FFmpeg example, so it should be safe.

This fixes issue #205.