Change game language, tabs, sorting and song menu without the need for a restart.
A restart used to be required in order to reflect changes in either the game language, the use of tabs, the sorting criterion or the type of song menu.
With these changes, all of the above can be changed and changes are reflected immediately.
For a language change, it is not even necessary to leave the current options screen.

This fixes #114 and #162.