Improve ebuild dependencies
2017.8.0 was not yet compatible with the 64 bit lua_Integer of Lua 5.3.
Therefore <dev-lang/lua-5.3. This has been fixed by 33fe22e3cfae1d67a74,
so 9999 does not need this.

We don't link directly to libpng, but we need SDL2_image to support PNG
and JPEG. Therefore the media-libs/libpng line should be deleted and
and the SDL2_image dependency should be changed to
media-libs/sdl2-image[png,jpeg] in both ebuilds.

Since 896390877bc3b0f8d6 portmidi can be used to play notes in the
built-in editors.