UImage: correctly disable alpha blending in FitImage
With the switch to SDL 2.0 SDL_SetAlpha(, 0, 255) was replaced
with SDL_SetSurfaceAlphaMod(, 255). This is incorrect as
SDL_SetSurfaceAlphaMod configures an additional factor to modify
the alpha values during blending while SDL_SetAlpha with second
parameter set to zero disables alpha blending. The correct way
to disable alpha blending with SDL 2.0 is to set the blend mode
of the source surface to SDL_BLENDMODE_NONE.

Without blending the pixel values are slightly different where
alpha is neither 255 nor 0 and now matches the unmodified images
that we already used when image dimensions were a power of two.

Fixes #456 because with blending disabled we completely avoid
the buggy NEON code added by Raspbian to SDL.