Further refactoring of the pitch detection algorithm.
Instead of defining constants for the lowest frequency (BaseToneFreq) and the halftone base (HalftoneBase), the only constant now required is the reference frequency of 440 Hz (A4), all other frequencies (and corresponding delays) for pitch analysis are derived from this smooth reference frequency.

The frequencies, the delays and the weights are currently stored in arrays which allows a verification of the pitch detection algorithms.

Currently, the previous AMDF functionality is in place via AnalyzePitch(PDA_AMDF), which can be switched to the circular average magnitude difference function approach via AnalyzePitch(PDA_CAMDF).

Commenting in lines 386-388 will output the detected tones (ToneAbs + corresponding frequencies) of both AMDF and CAMDF for comparison.