Use Lua library returned by pkg-config
Check each of the libraries returned by pkg-config if they contain
the Lua interpreter. This requires that the shared library is already
linked to all libraries it needs so that the other libraries returned
by pkg-config are important only for static linking. This approach
will work only for shared libraries.

Free Pascal doesn't really care if we pass 'liblua.so', 'liblua' or
'lua'. It will strip 'lib' from the beginning and (on Linux) '.so'
from the end if they exist before it writes the string to the linker
script.

If the library can't be determined, the configure script behaves like
before and uses the pkg-config module name as library name.