limit supported projectM version
In version 3.0.0 projectM started its move to legacy-less OpenGL. Without
any changes the graphical output is completely broken with current
projectM versions. Adding a glUseProgram(0) to RestoreOpenGLState fixes
the graphics we draw on top of the visualizations, but the visualizations
themself still look broken with some OpenGL drivers. It is also
questionable if our approach of reinitializing projectM's OpenGL state
from scratch for every frame will lead to acceptable performance if
shaders have to be compiled.