AppImage: no need to build python in CentOS 7
The python venv will create a symlink from python3 to python.
This causes scan_libs.py to be executed with that old python 3.6.8.
That version did not yet output UTF-8 in the C locale. Since
scan_libs.py uses non-ASCII to draw a beautiful tree, we get
an exception when we build in the default locale