Song Editor: change BPM value via F5
Changing the BPM in this way will cause a recalculation of all note timings (start + duration). This differs from adapting the BPM value via +/- which does not change any not timings. The use case for this is to correct wrong BPMs of songs that are otherwise synchronous. A note’s start beat is always rounded up (potentially starting a little later than what would be exact), a note’s duration is always rounded down (potentially lasting a little shorter than what would be exact).