Fix recalculation of GAP after setting timestamp of first note to zero.
If the start beat of the first note is not zero, pressing T in the editor shifts all notes so that the first note starts at zero and adjusts the GAP accordingly (good practice). However, there was an error in the calculation (CurrentSong.BPM[0].BPM is beats per 60.000 ms), which is fixed with this commit. Tested on several songs.