UVideo: free data of last decoded frame
While avcodec_receive_frame does call av_frame_unref on the frame we
pass to it, there is nobody who will free the data of the last decoded
frame for us. The fact that the codec won't unref the frame when it is
closed was documented at the refcounted_frames flag, where it said that
refcounted_frames is automatically enabled when avcodec_receive_frame
is used.