Properly fix memory leak in audio packet queue
The initial implementation of the packet queue never intended to
duplicate packets. The purpose of av_dup_packet is not to create a
copy of a packet that can be freed independently. What is does is
that it creates a managed copy of the data of a packet only if
av_read_frame returned a packet with unmanaged data. In cases where
the data returned by av_read_frame is already managed (i.e. it has
a destructor or is reference counted), av_dup_packet does nothing.
Unmanaged data has to be copied because it is valid only until the
next call to av_read_frame.

So when av_dup_packet vanished with FFmpeg 5.0, it was not correct
to replace it with av_packet_ref. Since with FFmpeg 5.0 all packets
returned by av_read_frame are reference counted, we have to do the
same av_dup_packet did for reference counted packets: nothing.