dldlls: don't use empty token and try GITHUB_TOKEN
Merge requests from other repositories will not have access to the
secret used for the ARTIFACT_ACCESS_TOKEN environment variable. The
variable will be set to an empty string. Using an empty token in the
Authorization header makes Github reject requests which would be
allowed without that header.

If the ARTIFACT_ACCESS_TOKEN is not available, try the GITHUB_TOKEN
to avoid the throttling of requests without authorization. It won't
work for downloading artifacts, though.